/account becomes the dashboard; setup moves to /account/settings
Two related complaints, one consolidated fix.

1. First login landed on the marketing homepage.
   The auth callback already redirects to /account, but if anything
   bumped the user back to '/' (clicking the logo, opening a bookmark,
   the magic-link link target rewriting) they ended up on the
   logged-out-looking pitch page.

   Fix: proxy.ts now redirects signed-in visitors hitting '/' to
   /account. The marketing homepage is for logged-out visitors only;
   signed-in users came back to USE the product, not read the pitch.

2. Clicking 'Dashboard' opened an interim-signup-looking setup page.
   /account rendered a stack of setup cards — mandate editor, buy-
   mandate editor, agent picker, visibility panel — with no overview
   of the actual portfolio (cash, holdings, trades). For a user with
   a working portfolio this felt like a perpetual onboarding screen.

   Fix: /account is now a router:
     * not signed in   → /login?next=/account
     * has portfolio   → /portfolios/<slug>      (the real dashboard)
     * no portfolio    → onboarding (CreatePortfolioForm)

   The setup cards moved to a new /account/settings route, reachable
   from the nav and from a 'Settings →' chip on the portfolio
   detail page header (owner-only).

3. Nav reshuffle.
   The redundant 'Portfolio' link (which redirected to /portfolios/<slug>
   via /account/portfolio) is now done by 'Dashboard' itself, so the
   third authed-only slot becomes 'Settings'. Final authed-only set:
   Dashboard | Watchlist | Settings.

   /account/portfolio (the explicit redirect route) stays for any
   bookmark / external link that might point at it.

Settings page is the same shape as the previous /account portfolio
view, minus the 'View portfolio page' chip (replaced with a '← Back
to dashboard' link) and with the 'Settings' page title.

Portfolio detail page (/portfolios/[slug]) gains a small 'Settings →'
chip next to the visibility toggle, owner-only. The chip uses the
same monospaced-pill style as the existing 'View portfolio page'
chip on the settings page header.
