+++ web/components/nav.tsx
@@ -13,14 +13,15 @@ const PUBLIC_LINKS: { href: string; label: string }[] = [
 ];
 
 // Links injected only when the visitor is signed in. Slotted in front of
-// the public set so the user's own surfaces ("Dashboard" / "Watchlist" /
-// "Portfolio") sit at the start of the nav row. "Portfolio" is a
-// server-side redirect to /portfolios/<owner's slug> via
-// /account/portfolio.
+// the public set so the user's own surfaces sit at the start of the nav
+// row. "Dashboard" lands on /account, which redirects to the user's
+// portfolio detail page (the actual overview). "Settings" hosts the
+// mandate / agent / visibility editors (moved off /account so the
+// dashboard isn't an interim signup-looking page).
 const AUTHED_LINKS: { href: string; label: string }[] = [
   { href: "/account", label: "Dashboard" },
   { href: "/account/watchlist", label: "Watchlist" },
-  { href: "/account/portfolio", label: "Portfolio" },
+  { href: "/account/settings", label: "Settings" },
 ];
 
 export default function Nav() {
